Solo dios es conciente de la pelea que se llevo aqui a cabo para poder compilar el codigo, bromas aparte todo parece funcional ahora pero hay que hacer mas pruebas para verificar que todo este bien, y porfavor tratar este codigo con cuidado que a la minima que tocas cualquier cosa todo deja de funcionar sin razon

+++ app/src/main/java/com/example/liftlog/model/Usuario.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "usuarios")
 data class Usuario(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0, // Actualizado a Long para consistencia
     val email: String,
     val password: String,
     val nombre: String,
